added put server route to handle updating of profile photo

+++ client/src/components/DeleteImage/deleteImage.js
@@ -2,10 +2,13 @@ import React, { Component } from "react";
 import API from "../../utils/API";
 import DropdownButton from 'react-bootstrap/DropdownButton'
 import Dropdown from 'react-bootstrap/Dropdown'
+import Image from 'react-bootstrap/Image';
+import Button from 'react-bootstrap/Button'
 import { MDBContainer } from 'mdbreact';
 import Card from 'react-bootstrap/Card'
 import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
+import './style.css'
 
 
 
@@ -14,12 +17,15 @@ import Col from 'react-bootstrap/Col'
 // const Auth = new AuthService();
 
 
+
 class DeleteImage extends Component {
   state = {
     NatureGallery: [],
     EngagementGallery: [],
     FoodGallery: [],
   }
+
+
 
   componentDidMount() {
     API
@@ -37,31 +43,58 @@ class DeleteImage extends Component {
 
   }
 
+
+
   handleDeleteNatureImage = (fileName) => {
     console.log(fileName)
-    prompt("Do you want to delete this image?") // pick back up here 
-    API
-    .deleteNatureImage(fileName)
-    .then(res => console.log(res))
-    .catch(err => console.log(err));
+    const valuePrompt = prompt("Do you want to delete this image? Type Yes to confirm") // pick back up here
+    if (valuePrompt === "Yes") {
+      API
+        .deleteNatureImage(fileName)
+        .then(res => {
+          console.log(res)
+          alert(`Image ${fileName} Deleted!`)
+          window.location.reload();
+        })
+        .catch(err => console.log(err));
+    } else {
+      alert("Delete Request cancelled")
+    }
   }
 
   handleDeleteEngagementImage = (fileName) => {
     console.log(fileName)
-    API
-    .deleteEngagementImage(fileName)
-    .then(res => console.log(res))
-    .catch(err => console.log(err));
+    const valuePrompt = prompt("Do you want to delete this image? Type Yes to confirm") // pick back up here
+    if (valuePrompt === "Yes") {
+      API
+        .deleteEngagementImage(fileName)
+        .then(res => {
+          console.log(res)
+          alert(`Image ${fileName} Deleted!`)
+          window.location.reload();
+        })
+        .catch(err => console.log(err));
+    } else {
+      alert("Delete Request cancelled")
+    }
   }
 
   handleDeleteFoodImage = (fileName) => {
     console.log(fileName)
-    API
-    .deleteFoodImage(fileName)
-    .then(res => console.log(res))
-    .catch(err => console.log(err));
+    const valuePrompt = prompt("Do you want to delete this image? Type Yes to confirm") // pick back up here
+    if (valuePrompt === "Yes") {
+      API
+        .deleteFoodImage(fileName)
+        .then(res => {
+          console.log(res)
+          alert(`Image ${fileName} Deleted!`)
+          window.location.reload();
+        })
+        .catch(err => console.log(err));
+    } else {
+      alert("Delete Request cancelled")
+    }
   }
-
 
 
 
@@ -69,50 +102,95 @@ class DeleteImage extends Component {
   render() {
     return (
       <div className="container removeImage">
+        <MDBContainer>
 
+          <hr />
 
-        <hr />
+          <h3>Remove an Image</h3>
+          <Row>
+            <Col sm={4}>
 
-        <h3>Remove an Image</h3>
-        <div>
-
-          {['down'].map(direction => (
-            <DropdownButton
-              drop={direction}
-              variant="secondary"
-              title={` Drop ${direction} `}
-              id={`dropdown-button-drop-${direction}`}
-              key={direction}
-            >
-              <Dropdown.Item>Nature Gallery</Dropdown.Item>
-              <Dropdown.Divider />
               {this.state.NatureGallery.map(image => (
-                <Dropdown.Item onClick={() => this.handleDeleteNatureImage(image.fileName)}>
-                  {image.fileName}
-                </Dropdown.Item>
+                <Card className="deleteCard" border="secondary" style={{ width: '18rem', textAlign: 'Center' }}>
+                  <Image
+
+                    className="ImageHolder"
+                    src={image.imageURL}
+                    thumbnail
+                    height="200"
+                    width="200"
+
+                  />
+                  <Card.Body>
+                    <Card.Text>
+                      <p>{image.fileName}</p>
+                    </Card.Text>
+                    <Button
+                      variant="outline-primary"
+                      type="submit"
+                      className="btn btn-outline-primary  pt-2"
+                      onClick={() => this.handleDeleteNatureImage(image.fileName)}>Delete Image</Button>
+
+                  </Card.Body>
+                </Card>
               ))}
-              <Dropdown.Divider />
-              <Dropdown.Item>Engagement Gallery</Dropdown.Item>
-              <Dropdown.Divider />
+
+            </Col>
+            <Col sm={4}>
               {this.state.EngagementGallery.map(image => (
-                <Dropdown.Item  onClick={() => this.handleDeleteEngagementImage(image.fileName)}>
-                >{image.fileName}
-                </Dropdown.Item>
+                <Card className="deleteCard" border="secondary" style={{ width: '18rem', textAlign: 'Center' }}>
+                  <Image
+                    
+                    className="ImageHolder"
+                    src={image.imageURL}
+                    thumbnail
+                    height="200"
+                    width="200"
+
+                  />
+                  <Card.Body>
+                    <Card.Text>
+                    <p>{image.fileName}</p>
+                </Card.Text>
+                    <Button
+                      variant="outline-primary"
+                      type="submit"
+                      className="btn btn-outline-primary  pt-2"
+                      onClick={() => this.handleDeleteEngagementImage(image.fileName)}>Delete Image</Button>
+
+                  </Card.Body>
+                </Card>
               ))}
-              <Dropdown.Item>Food Gallery</Dropdown.Item>
-              <Dropdown.Divider />
+            </Col>
+            <Col sm={4}>
               {this.state.FoodGallery.map(image => (
-                <Dropdown.Item onClick={() => this.handleDeleteFoodImage(image.fileName)}>
-                {image.fileName}
-                </Dropdown.Item>
+                <Card className="deleteCard" border="secondary" style={{ width: '18rem', textAlign: 'Center' }}>
+                  <Image
+
+                    className="ImageHolder"
+                    src={image.imageURL}
+                    thumbnail
+                    height="200"
+                    width="200"
+                  
+
+                  />
+                  <Card.Body>
+                    <Card.Text>
+                    <p>{image.fileName}</p>
+                </Card.Text>
+                    <Button
+                      variant="outline-primary"
+                      type="submit"
+                      className="btn btn-outline-primary  pt-2"
+                      onClick={() => this.handleDeleteFoodImage(image.fileName)}>Delete Image</Button>
+
+                  </Card.Body>
+                </Card>
               ))}
-            </DropdownButton>
-          ))}
-
-
-
-
-        </div>
+            </Col>
+          </Row>
+        </MDBContainer>
       </div>
     )
   }
@@ -122,18 +200,3 @@ class DeleteImage extends Component {
 export default DeleteImage;
 
 
-{/* <Button  
-type="submit"
-className="btn btn-outline-primary btn-block pt-2"
-onClick={this.handleSubmitNature}
-variant="outline-success">Nature Gallery</Button>
-
-<Button variant="outline-danger" 
-type="submit"
-className="btn btn-outline-primary btn-block pt-2"
-onClick={this.handleSubmitEngagement}>Engagement Gallery</Button>
-
-<Button 
-variant="outline-warning" type="submit"
-className="btn btn-outline-primary btn-block pt-2"
-onClick={this.handleSubmitFood}>Food Gallery</Button */}
